Add -ffp-contract=off for iOS and MacOS ARM builds
Apparently, Apple's ARM compilers default to using FMA for expressions
like a * b - c * d, which causes undesirable results in Free42 and
Plus42. Specifically, (pi-pi*i)*(pi-pi*i) was producing a result with
nonzero real part in the binary builds.
Note that the x86_64 builds, iOS and MacOS, were not affected, and
neither were the Linux and Windows builds. (I haven't tested this on
Android.)